Revert "Check point 1"
