+++ .gitignore
@@ -29,6 +29,3 @@ target
 # Eclipse
 *.classpath
 *.project
-
-# Questions
-/question*.txt
