+++ core/src/main/java/com/es/core/cart/Cart.java
@@ -1,0 +1,4 @@
+package com.es.core.cart;
+
+public class Cart {
+}
