+++ core/src/main/java/com/es/core/cart/CartService.java
@@ -1,0 +1,19 @@
+package com.es.core.cart;
+
+import java.util.Map;
+
+public interface CartService {
+
+    Cart getCart();
+
+    void addPhone(Long phoneId, Long quantity);
+
+    /**
+     * @param items
+     * key: {@link com.es.core.model.phone.Phone#id}
+     * value: quantity
+     */
+    void update(Map<Long, Long> items);
+
+    void remove(Long phoneId);
+}
