--- core/src/main/java/com/es/core/model/DAO/phone/PhoneDao.java
@@ -1,12 +1,0 @@
-package com.es.core.model.DAO.phone;
-
-import com.es.core.model.entity.phone.Phone;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface PhoneDao {
-    Optional<Phone> get(Long key);
-    void save(Phone phone);
-    List<Phone> findAll(int offset, int limit);
-}
