--- core/src/main/java/com/es/core/model/DAO/phone/PhoneFieldsConstantsController.java
@@ -1,30 +1,0 @@
-package com.es.core.model.DAO.phone;
-
-public class PhoneFieldsConstantsController {
-    public static final String phone_id = "id";
-    public static final String brand_unique = "brand";
-    public static final String model_unique = "model";
-    public static final String price = "price";
-    public static final String displaySizeInches = "displaySizeInches";
-    public static final String weightGr = "weightGr";
-    public static final String lengthMm = "lengthMm";
-    public static final String widthMm = "widthMm";
-    public static final String heightMm = "heightMm";
-    public static final String announced = "announced";
-    public static final String deviceType = "deviceType";
-    public static final String os = "os";
-    public static final String displayResolution = "displayResolution";
-    public static final String pixelDensity = "pixelDensity";
-    public static final String displayTechnology = "displayTechnology";
-    public static final String backCameraMegapixels = "backCameraMegapixels";
-    public static final String frontCameraMegapixels = "frontCameraMegapixels";
-    public static final String ramGb = "ramGb";
-    public static final String internalStorageGb = "internalStorageGb";
-    public static final String batteryCapacityMah = "batteryCapacityMah";
-    public static final String talkTimeHours = "talkTimeHours";
-    public static final String standByTimeHours = "standByTimeHours";
-    public static final String bluetooth = "bluetooth";
-    public static final String positioning = "positioning";
-    public static final String imageUrl = "imageUrl";
-    public static final String description = "description";
-}
