--- core/src/main/java/com/es/core/model/entity/order/OrderStatus.java
@@ -1,5 +1,0 @@
-package com.es.core.model.entity.order;
-
-public enum OrderStatus {
-    NEW, DELIVERED, REJECTED
-}
