--- core/src/main/java/com/es/core/model/entity/phone/Color.java
@@ -1,54 +1,0 @@
-package com.es.core.model.entity.phone;
-
-import java.util.Objects;
-
-public class Color {
-    private Long id;
-    private String code;
-
-    public Color() {
-    }
-
-    public Color(Long id, String code) {
-        this.id = id;
-        this.code = code;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getCode() {
-        return code;
-    }
-
-    public void setCode(final String code) {
-        this.code = code;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) return true;
-        if (o == null || getClass() != o.getClass()) return false;
-        Color color = (Color) o;
-        return Objects.equals(id, color.id) &&
-                Objects.equals(code, color.code);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, code);
-    }
-
-    @Override
-    public String toString() {
-        return "Color{" +
-                "id=" + id +
-                ", code='" + code + '\'' +
-                '}';
-    }
-}
