+++ core/src/main/java/com/es/core/model/order/OrderStatus.java
@@ -1,0 +1,5 @@
+package com.es.core.model.order;
+
+public enum OrderStatus {
+    NEW, DELIVERED, REJECTED
+}
