+++ core/src/main/java/com/es/core/model/phone/JdbcPhoneDao.java
@@ -1,0 +1,27 @@
+package com.es.core.model.phone;
+
+import org.springframework.jdbc.core.BeanPropertyRowMapper;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.stereotype.Component;
+
+import javax.annotation.Resource;
+import java.util.List;
+import java.util.Optional;
+
+@Component
+public class JdbcPhoneDao implements PhoneDao{
+    @Resource
+    private JdbcTemplate jdbcTemplate;
+
+    public Optional<Phone> get(final Long key) {
+        throw new UnsupportedOperationException("TODO");
+    }
+
+    public void save(final Phone phone) {
+        throw new UnsupportedOperationException("TODO");
+    }
+
+    public List<Phone> findAll(int offset, int limit) {
+        return jdbcTemplate.query("select * from phones offset " + offset + " limit " + limit, new BeanPropertyRowMapper(Phone.class));
+    }
+}
