+++ core/src/main/java/com/es/core/model/phone/PhoneDao.java
@@ -1,0 +1,10 @@
+package com.es.core.model.phone;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface PhoneDao {
+    Optional<Phone> get(Long key);
+    void save(Phone phone);
+    List<Phone> findAll(int offset, int limit);
+}
