+++ core/src/main/java/com/es/core/model/phone/Stock.java
@@ -1,0 +1,31 @@
+package com.es.core.model.phone;
+
+public class Stock {
+    private Phone phone;
+    private Integer stock;
+    private Integer reserved;
+
+    public Phone getPhone() {
+        return phone;
+    }
+
+    public void setPhone(Phone phone) {
+        this.phone = phone;
+    }
+
+    public Integer getStock() {
+        return stock;
+    }
+
+    public void setStock(Integer stock) {
+        this.stock = stock;
+    }
+
+    public Integer getReserved() {
+        return reserved;
+    }
+
+    public void setReserved(Integer reserved) {
+        this.reserved = reserved;
+    }
+}
