+++ core/src/main/java/com/es/core/order/OrderService.java
@@ -1,0 +1,9 @@
+package com.es.core.order;
+
+import com.es.core.cart.Cart;
+import com.es.core.model.order.Order;
+
+public interface OrderService {
+    Order createOrder(Cart cart);
+    void placeOrder(Order order) throws OutOfStockException;
+}
