+++ core/src/main/java/com/es/core/order/OrderServiceImpl.java
@@ -1,0 +1,18 @@
+package com.es.core.order;
+
+import com.es.core.cart.Cart;
+import com.es.core.model.order.Order;
+import org.springframework.stereotype.Service;
+
+@Service
+public class OrderServiceImpl implements OrderService {
+    @Override
+    public Order createOrder(Cart cart) {
+        throw new UnsupportedOperationException("TODO");
+    }
+
+    @Override
+    public void placeOrder(Order order) throws OutOfStockException {
+        throw new UnsupportedOperationException("TODO");
+    }
+}
