+++ core/src/main/java/com/es/core/order/OutOfStockException.java
@@ -1,0 +1,4 @@
+package com.es.core.order;
+
+public class OutOfStockException extends Exception {
+}
