--- core/src/main/java/com/es/core/service/cart/Cart.java
@@ -1,4 +1,0 @@
-package com.es.core.service.cart;
-
-public class Cart {
-}
