--- core/src/main/java/com/es/core/service/order/OrderService.java
@@ -1,9 +1,0 @@
-package com.es.core.service.order;
-
-import com.es.core.service.cart.Cart;
-import com.es.core.model.entity.order.Order;
-
-public interface OrderService {
-    Order createOrder(Cart cart);
-    void placeOrder(Order order) throws OutOfStockException;
-}
