--- core/src/main/java/com/es/core/service/order/OrderServiceImpl.java
@@ -1,18 +1,0 @@
-package com.es.core.service.order;
-
-import com.es.core.service.cart.Cart;
-import com.es.core.model.entity.order.Order;
-import org.springframework.stereotype.Service;
-
-@Service
-public class OrderServiceImpl implements OrderService {
-    @Override
-    public Order createOrder(Cart cart) {
-        throw new UnsupportedOperationException("TODO");
-    }
-
-    @Override
-    public void placeOrder(Order order) throws OutOfStockException {
-        throw new UnsupportedOperationException("TODO");
-    }
-}
