--- core/src/main/java/com/es/core/service/order/OutOfStockException.java
@@ -1,4 +1,0 @@
-package com.es.core.service.order;
-
-public class OutOfStockException extends Exception {
-}
