--- core/src/main/resources/config/application.properties
@@ -1,6 +1,0 @@
-db.driver=org.h2.Driver
-db.url=jdbc:h2:~/phoneshop/db/h2
-db.user=sa
-db.password=
-db.insertSchema=true
-db.insertDemodata=true
