+++ core/src/main/resources/db/schema.sql
