--- core/src/main/test/resources/config/test-config.properties
@@ -1,4 +1,0 @@
-db.driver=org.h2.Driver
-db.url=jdbc:h2:~/phoneshop/db/h2
-db.user=sa
-db.password=
