+++ demodata/pom.xml
@@ -7,6 +7,7 @@
     <version>1.0-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
+  <groupId>com.es.phoneshop</groupId>
   <artifactId>phoneshop-demodata</artifactId>
   <packaging>jar</packaging>
   <version>1.0-SNAPSHOT</version>
