--- run.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-mvn tomcat7:run
