+++ web/src/main/java/com/es/phoneshop/web/controller/pages/CartPageController.java
@@ -1,6 +1,6 @@
 package com.es.phoneshop.web.controller.pages;
 
-import com.es.core.service.cart.CartService;
+import com.es.core.cart.CartService;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
