+++ web/src/main/webapp/WEB-INF/conf/application.properties
@@ -1,6 +1,6 @@
 # TODO
 # db.driver=
-# db.url=
+# db.ulr=
 # db.user=
 # db.password=
 # db.insertSchema=true/false
